modify: modify code to improve code quality
